Revert "Move into keypair package"
This reverts commit b4372f540704e078455f19fbc34f0ca91f88a681.
