+++ keypairgen/keypairgen.go
@@ -1,4 +1,6 @@
-package keypair
+package keypairgen
+
+import "github.com/stellar/go/keypair"
 
 // Generator generates new keys with the underlying source. The underlying
 // source defaults to the RandomSource if not specified.
@@ -14,20 +16,20 @@ func (g *Generator) getSource() Source {
 }
 
 // Generate returns a new key using the underlying source.
-func (g *Generator) Generate() (*Full, error) {
+func (g *Generator) Generate() (*keypair.Full, error) {
 	return g.getSource().Generate()
 }
 
 // Source provides keys.
 type Source interface {
-	Generate() (*Full, error)
+	Generate() (*keypair.Full, error)
 }
 
 // RandomSource provides new keys that are randomly generated using the
-// Random function.
+// keypair.Random function.
 type RandomSource struct{}
 
 // Generated returns a new key using keypair.Random.
-func (RandomSource) Generate() (*Full, error) {
-	return Random()
+func (RandomSource) Generate() (*keypair.Full, error) {
+	return keypair.Random()
 }
